workflows/train/transformers/glue/requirements.txt
git+git://github.com/huggingface/transformers.git
datasets
azureml-sdk
torch
scipy
sklearn